+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] 2018-12-07
+### Added
+- Support for Ruby version 2.6.0-rc1
+
+### Fixes
+- Fixed issue with tar reader not detecting directory within tar archive
+
 ## [2.1.0] 2018-07-08
 ### Added
 - TODO: When a new global Ruby version is installed, the latest Mineshaft gem will be installed as well